+++ IrrigationApi/Irrigation.Api/Migrations/IrrigationDataContextModelSnapshot.cs
@@ -24,12 +24,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.Area", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("area_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(max)")
@@ -50,12 +50,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.Notification", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("notification_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(max)")
@@ -65,8 +65,8 @@ namespace Irrigation.Api.Migrations
                         .HasColumnType("datetime2")
                         .HasColumnName("timestamp");
 
-                    b.Property<long?>("sensor_id")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("sensor_id")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -96,12 +96,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.Schedule", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("schedule_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("EndTime")
                         .HasColumnType("datetime2")
@@ -118,12 +118,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.Sensor", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("sensor_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Location")
                         .HasColumnType("nvarchar(max)")
@@ -133,8 +133,8 @@ namespace Irrigation.Api.Migrations
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("type");
 
-                    b.Property<long?>("area_id")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("area_id")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -145,12 +145,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.User", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("user_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<bool>("Active")
                         .HasColumnType("BIT")
@@ -185,12 +185,12 @@ namespace Irrigation.Api.Migrations
 
             modelBuilder.Entity("Irrigation.Core.Models.Weather", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasColumnName("weather_id");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(max)")
@@ -208,8 +208,8 @@ namespace Irrigation.Api.Migrations
                         .HasColumnType("datetime2")
                         .HasColumnName("timestamp");
 
-                    b.Property<long?>("sensor_id")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("sensor_id")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -223,8 +223,8 @@ namespace Irrigation.Api.Migrations
                     b.Property<int>("role_id")
                         .HasColumnType("int");
 
-                    b.Property<long>("user_id")
-                        .HasColumnType("bigint");
+                    b.Property<int>("user_id")
+                        .HasColumnType("int");
 
                     b.HasKey("role_id", "user_id");
